The New Era Update
The era of dynamic bone been ended. As Avatar Dynamic have recently been enabled for Live, it has been decided to move Dynamic Bone to the Legacy Components. This change was a vote taken on the Discord. Thanks for the amazing years, Dynamic Bone. Long Live Avatar Dynamics and PhysBones.